+++ index.d.ts
@@ -1,8 +1,9 @@
-import { TInputOptions } from './src/types/options/TInputOptions';
 import { TDictionary } from './src/types/TDictionary';
+import { TInputOptions } from './src/types/options/TInputOptions';
+import { TObfuscationResultsObject } from './src/types/TObfuscationResultsObject';
+import { TOptionsPreset } from './src/types/options/TOptionsPreset';
 
 import { IObfuscatedCode } from './src/interfaces/source-code/IObfuscatedCode';
-import { TObfuscationResultsObject } from './src/types/TObfuscationResultsObject';
 
 export type ObfuscatorOptions = TInputOptions;
 
@@ -24,6 +25,12 @@ export function obfuscateMultiple <TSourceCodesObject extends TDictionary<string
     sourceCodesObject: TSourceCodesObject,
     inputOptions?: TInputOptions
 ): TObfuscationResultsObject<TSourceCodesObject>;
+
+/**
+ * @param {TOptionsPreset} optionsPreset
+ * @returns {TInputOptions}
+ */
+export function getOptionsByPreset (optionsPreset: TOptionsPreset): TInputOptions;
 
 /**
  * @type {string}