+++ src/types/options/TOptionsPreset.ts
@@ -1,0 +1,5 @@
+import { TypeFromEnum } from '@gradecam/tsenum';
+
+import { OptionsPreset } from '../../enums/options/presets/OptionsPreset';
+
+export type TOptionsPreset = TypeFromEnum<typeof OptionsPreset>;