+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 Change Log
 
+v2.1.0
+---
+* **New API:** `getOptionsByPreset` allows to get options for the passed options preset name 
+
 v2.0.0
 ---
-* **Breaking change:** `stringArrayEncoding` option now accepts an array of encodings. Each string will be randomly encoded with passed encoding.
+* **Breaking change:** `stringArrayEncoding` option now accepts an array of encodings. Each string will be randomly encoded with passed encoding
 
 v1.12.1
 ---